convert tabs to spaces
Using tabs in the ui breaks the ratatui diff and leads to rendering artifacts.
This converts tabs to spaces with a fixed tab-width of 4. Ideally, the
tab-width should be configurable, but until then 4 might be a reasonable
compromise.
(I'm assuming that few people who indent with tabs are using a larger
tab-width.)