log tab: bind `s` to squash changes from @ into the selected change
For now, this just squashes everything into the selected commit,
which is quite a large hammer, but still useful when working on
in a mega merge workflow and pushing down a whole set of changes
into a new parent.

It would be useful to add a way to do an interactive squashing, but
the current support to run jj command only allows non-interactive ones,
so we'll have to add that later.