Fix terminal on panic (#93)
* Remove unnecessary call to show_cursor()

We don't explicitly hide the cursor, so we don't have to show it either.

* Make restore_terminal() not depend on a Terminal instance

The way the code is written, it's enforced by the type system, that the
backend uses Stdout, so we can just use Stdout directly. This makes it
easier to call restore_terminal() from other places, like e.g. a panic
handler to restore the terminal on panics and not leave the terminal
garbled and in a messed up state.

* Install a panic handler to restore the terminal state on panic

Otherwise, the terminal is left all messed up on panics.