fix: jj 0.33.0 and Rust 1.89.0 compatibility (#159)
* chore: Collapse nested if blocks to make Clippy happy

* chore: Don't mix hidden and elided lifetimes

That's a new warning in Rust 1.89.0.

* fix: jj 0.33.0 compatibility

--config-toml has been removed so we need to use --config instead, which
has a bit stricter format requirements. No leading/trailing new lines,
and no spaces around the equal sign.

It can also only take a single option at a time, so we need to change
jj_config_toml from a String to a Vec<String>.

And for obslog/evolog, the template format has changed, so we need to
update that as well.